Menu

+++ src/main/java/com/uplog/uplog/domain/post/application/PostService.java
@@ -1,7 +1,10 @@
 package com.uplog.uplog.domain.post.application;
 
+import com.uplog.uplog.domain.comment.dao.CommentRepository;
+import com.uplog.uplog.domain.like.dao.PostLikeRepository;
 import com.uplog.uplog.domain.member.dao.MemberRepository;
 import com.uplog.uplog.domain.member.model.Member;
+import com.uplog.uplog.domain.member.model.Position;
 import com.uplog.uplog.domain.menu.dao.MenuRepository;
 import com.uplog.uplog.domain.menu.model.Menu;
 import com.uplog.uplog.domain.post.dao.PostRepository;
@@ -15,6 +18,7 @@ import com.uplog.uplog.domain.project.model.Project;
 import com.uplog.uplog.domain.task.exception.NotFoundTaskByIdException;
 import com.uplog.uplog.global.exception.AuthorityException;
 import com.uplog.uplog.global.exception.NotFoundIdException;
+import com.uplog.uplog.global.method.AuthorizedMethod;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -32,26 +36,33 @@ public class PostService {
     private final ProductRepository productRepository;
     private final ProjectRepository projectRepository;
     private final MenuRepository menuRepository;
+    private final AuthorizedMethod authorizedMethod;
+    private final PostLikeRepository postLikeRepository;
+    private final CommentRepository commentRepository;
 
     /*
     Create
      */
     @Transactional
-    public Post createPost(Long id, CreatePostRequest createPostRequest) {
+    public PostInfoDTO createPost(Long id, CreatePostRequest createPostRequest) {
         Member author = memberRepository.findMemberById(id)
-                .orElseThrow(() -> new RuntimeException("Member not found"));
+                .orElseThrow(() -> new NotFoundIdException("해당 멤버는 존재하지 않습니다."));
+
+        //현재 프로젝트 팀 내에 존재하는 멤버,기업이 아닌 회원,클라이언트가 아닌 멤버 확인
+        authorizedMethod.CreatePostTaskValidateByMemberId(author);
 
         Menu menu = menuRepository.findById(createPostRequest.getMenuId())
-                .orElseThrow(() -> new RuntimeException("Menu not found"));
+                .orElseThrow(() -> new NotFoundIdException("해당 메뉴는 존재하지 않습니다."));
 
         Project project = projectRepository.findById(createPostRequest.getProjectId())
-                .orElseThrow(() -> new RuntimeException("Project not found"));
+                .orElseThrow(() -> new NotFoundIdException("해당 프로젝트는 존재하지 않습니다."));
 
-        Product product = productRepository.findById(createPostRequest.getPorductId())
-                .orElseThrow(() -> new RuntimeException("Product not found"));
+        Product product = productRepository.findById(createPostRequest.getProductId())
+                .orElseThrow(() -> new NotFoundIdException("해당 제품은 존재하지 않습니다."));
 
 
-       // Post post = createPostRequest.toEntity(author, menu, product, project);
+
+        // Post post = createPostRequest.toEntity(author, menu, product, project);
         PostType postType = PostType.DEFAULT; // 기본값으로 설정
 
         String requestType = createPostRequest.getPostType();
@@ -66,11 +77,11 @@ public class PostService {
             }
         }
 
-
-        Post post=createPostRequest.toEntity(author,menu, product, project,postType);
+        //TODO 해당 프로젝트사람들이면
+        Post post = createPostRequest.toEntity(author, menu, product, project, postType);
         postRepository.save(post);
 
-        return post;
+        return toPostInfoDTO(post);
 
     }
 
@@ -79,9 +90,16 @@ public class PostService {
     Delete
      */
     @Transactional
-    public void deletePost(Long id) {
+    public String deletePost(Long id,Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundIdException::new);
-        postRepository.delete(post);
+
+        if(post.getAuthor().getId().equals(currentUserId)){
+            postRepository.delete(post);
+            return "delete 완료";
+        }
+        else{
+            throw new AuthorityException("작성자와 일치하지 않아 삭제 권한이 없습니다.");
+        }
     }
 
     /*
@@ -89,34 +107,34 @@ public class PostService {
     */
     //TODO update 권한 설정해야,==으로 바꾸기
     @Transactional
-    public Post updatePostTitle(Long id, UpdatePostTitleRequest updatePostTitleRequest,Long currentUserId) {
+    public PostInfoDTO updatePostTitle(Long id, UpdatePostTitleRequest updatePostTitleRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
-          if(post.getAuthor().getId().equals(currentUserId)){
-              post.updatePostTitle(updatePostTitleRequest.getUpdateTitle());
-              return post;
+        if(post.getAuthor().getId().equals(currentUserId)){
+            post.updatePostTitle(updatePostTitleRequest.getUpdateTitle());
+            return toPostInfoDTO(post);
         }
         else{
-            throw new AuthorityException();
+            throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
         }
 
     }
 
     @Transactional
-    public Post updatePostContent(Long id, UpdatePostContentRequest updatePostContentRequest,Long currentUserId) {
+    public PostInfoDTO updatePostContent(Long id, UpdatePostContentRequest updatePostContentRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         if(post.getAuthor().getId().equals(currentUserId)){
             post.updatePostContent(updatePostContentRequest.getUpdateContent());
-            return post;
+            return toPostInfoDTO(post);
         }
         else{
-            throw new AuthorityException();
+            throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
         }
 
     }
 
     //TODO Enum수정
     @Transactional
-    public Post updatePostType(Long id, UpdatePostTypeRequest updatePostTypeRequest,Long currentUserId) {
+    public PostInfoDTO updatePostType(Long id, UpdatePostTypeRequest updatePostTypeRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         PostType updatepostType = PostType.DEFAULT; // 기본값으로 설정
 
@@ -133,25 +151,25 @@ public class PostService {
                 }
             }
             post.updatePostType(updatepostType);
-            return post;
+            return toPostInfoDTO(post);
         }
         else{
-            throw new AuthorityException();
+            throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
         }
 
     }
 
     @Transactional
-    public Post updatePostMenu(Long id, UpdatePostMenuRequest updatePostMenuRequest,Long currentUserId) {
+    public PostInfoDTO updatePostMenu(Long id, UpdatePostMenuRequest updatePostMenuRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         Menu menu = menuRepository.findById(updatePostMenuRequest.getUpdateMenuId())
-                .orElseThrow(() -> new RuntimeException("Menu not found"));
+                .orElseThrow(() -> new NotFoundIdException("해당 메뉴는 존재하지 않습니다."));
         if(post.getAuthor().getId().equals(currentUserId)){
             post.updatePostMenu(menu);
-            return post;
+            return toPostInfoDTO(post);
         }
         else{
-            throw new AuthorityException();
+            throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
         }
 
     }
@@ -159,46 +177,67 @@ public class PostService {
 
     //TODO 이건 나중에 제품 수정할때 같이 불러야하는 서비스
     @Transactional
-    public Post updateProductName(Long id, UpdatePostProductRequest updatePostProductRequest,Long currentUserId) {
+    public PostInfoDTO updateProductName(Long id, UpdatePostProductRequest updatePostProductRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
-        if(post.getAuthor().getId().equals(currentUserId)){
-            post.updatePostProductName(updatePostProductRequest.getUpdateProductName());
-            return post;
-        }
-        else{
-            throw new AuthorityException();
-        }
+        post.updatePostProductName(updatePostProductRequest.getUpdateProductName());
+        return toPostInfoDTO(post);
+
 
     }
 
     //TODO 이건 나중에 프로젝트 수정할때 같이 불러야하는 서비스임
-
     @Transactional
-    public Post updateVersion(Long id, UpdatePostVersionRequest updatePostVersionRequest,Long currentUserId) {
+    public PostInfoDTO updateVersion(Long id, UpdatePostVersionRequest updatePostVersionRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
-        if(post.getAuthor().getId().equals(currentUserId)){
-            post.updatePostVersion(updatePostVersionRequest.getUpdateVersion());
-            return post;
-        }
-        else{
-            throw new AuthorityException();
-        }
+        post.updatePostVersion(updatePostVersionRequest.getUpdateVersion());
+        return toPostInfoDTO(post);
 
     }
 
     /*
     Get
      */
-    @Transactional
-    public List<PostInfoDTO> getPostByMenu(Long menuId){
+
+    @Transactional(readOnly = true)
+    public PostInfoDTO findById(Long Id){
+        Post post=postRepository.findById(Id).orElseThrow(NotFoundIdException::new);;
+        return toPostInfoDTO(post);
+    }
+    @Transactional(readOnly = true)
+    public List<PostInfoDTO> findPostInfoByMenuId(Long menuId){
         List<Post> postList=postRepository.findByMenuId(menuId);
         List<PostInfoDTO> postInfoDTOs=new ArrayList<>();
         for(Post post:postList){
-            PostInfoDTO postInfoDTO=post.toPostInfoDTO();
+            PostInfoDTO postInfoDTO=toPostInfoDTO(post);
             postInfoDTOs.add(postInfoDTO);
         }
         return postInfoDTOs;
 
+    }
+    //이건 안쓰는거->메뉴에서 씀
+    public List<Post> findPostsByMenuId(Long menuId) {
+        List<Post> postList = postRepository.findByMenuId(menuId);
+        return postList;
+    }
+
+    public PostInfoDTO toPostInfoDTO(Post post) {
+        int likeCount = postLikeRepository.countByPostId(post.getId());
+        int commentCount = commentRepository.countByPostId(post.getId());
+
+        return PostInfoDTO.builder()
+                .id(post.getId())
+                .title(post.getTitle())
+                .authorInfoDTO(post.getAuthor().powerMemberInfoDTO())
+                .menuId(post.getMenu().getId())
+                .menuName(post.getMenu().getMenuName())
+                .productName(post.getProductName())
+                .projectName(post.getVersion())
+                .postType(post.getPostType())
+                .content(post.getContent())
+                .createTime(post.getCreateTime())
+                .likeCount(likeCount)
+                .commentCount(commentCount)
+                .build();
     }
